Added feature to print release notes (#8)
It is sometimes useful to also see the release notes
related to project's packaged releases, not just the
commit SHAs and description. This commit adds the
'--release-notes' flag. When specified, osa-differ
will also print the release notes of the OSA packaged
releases between the two commits given.

NOTE: This will only print the release notes for the
OpenStack-Ansible project, not the upstream OpenStack
projects.