RE-1190 Fix wrong branch used
FETCH_HEAD is a git ref that points to the last thing that was fetched.
This is useful when a single ref has been fetched, however it is
unreliable when multiple refs have been fetched.

Before this commit, a reset --hard FETCH_HEAD was issued after fetching.
This will have moved the currently checkout branch to
one of the fetched refs, which one is non deterministic.
This commit removes the erroneous checkout.

However even without a rogue reset, there may be old local branches from
previous runs of osa_differ, so this commit also removes local branches
before checking out a branch of the same name. This ensures git will
use it's logic to find the upstream branch and create a local branch
from that.

Previously reset was used, which doesn't have the same logic for
finding matching remote branches, so this commit switches to checkout.