RE-1405 Fix 'branch is not fully merged' error
In 35549f4 a change in the checkout mechanism was
implemented which involved deleting the branch, but
in some cases the branch will not delete giving an
error like:

stderr: 'error: The branch 'RE-1405-master' is not fully merged.
If you are sure you want to delete it, run 'git branch -D RE-1405-master'.'

This patch implements the extra parameter to ensure
that the deletion is forced.