Add ability to override dependency versions
This change extends osa-differ to allow a version mapping to be
specified for a project or role. For example, if the project foo is
pinned to the version 1.0.0 in OpenStack-Ansible and the project removes
the tag and replaces it with v1.0.0, without this change the command
will fail. To address this, this change would allow the following the be
specified `--version-mapping 'foo;1.0.0:v1.0.0'`. A new commit-ish can
be proposed to override the one specified in OpenStack-Ansible to allow
osa-diff to work with versions of OpenStack-Ansible that no longer
specify valid project/role versions.

Version updated to 0.3.10 in preparation for a new release.

JIRA: RE-2208